add Order, improvements

+++ src/main/java/ru/yandex/practicum/dto/ItemDto.java
@@ -2,13 +2,15 @@ package ru.yandex.practicum.dto;
 
 import ru.yandex.practicum.model.Item;
 
+import java.math.BigDecimal;
+
 /**
  * DTO for {@link Item}
  */
 public record ItemDto(long id,
                       String title,
                       String description,
-                      long price,
+                      BigDecimal price,
                       String imgPath,
                       Integer count) {
 }
